created api routes for poll and recipes; Completed recipes create route with tags, ingredients and ratings

+++ src/main/java/com/example/ricettario/entities/RecipeIngredientId.java
@@ -7,13 +7,13 @@ import java.util.Objects;
 @Embeddable
 public class RecipeIngredientId implements Serializable {
 
-    private Long recipeId;
-    private Long ingredientId;
+    private int recipeId;
+    private int ingredientId;
 
     public RecipeIngredientId() {
     }
 
-    public RecipeIngredientId(Long recipeId, Long ingredientId) {
+    public RecipeIngredientId(int recipeId, int ingredientId) {
         this.recipeId = recipeId;
         this.ingredientId = ingredientId;
     }
